+++ test/integration/default/default_test.rb
@@ -23,3 +23,17 @@ describe file('/etc/mongod.conf') do
   it { should be_readable }
   its('content') { should match 'dbPath' }
 end
+
+# optimization for mongo client
+describe file('/etc/init.d/disable-transparent-hugepages') do
+  it { should exist }
+
+  describe command('chkconfig --list') do
+    its('stdout') { should match(/disable-transparent-hugepage /) }
+  end
+end
+
+# test for user authentication
+describe command('mongo') do
+  its('stdout') { should_not match(/WARNING: Access control/) }
+end